+++ interface/about_dialog.py
@@ -1,21 +1,23 @@
 import os
-import PyQt4.QtCore as QtCore
-import PyQt4.QtGui as QtGui
 
+from PyQt5 import QtCore, QtGui, QtWidgets
 from designer import about_dialog_qt
 
 
-class AboutDialog(QtGui.QDialog, about_dialog_qt.Ui_Dialog):
+class AboutDialog(QtWidgets.QDialog, about_dialog_qt.Ui_Dialog):
     def __init__(self, parent=None):
         super(AboutDialog, self).__init__(parent)
 
         # Setup the about dialog box
         self.setupUi(self)
 
+        # Set version number
+        self.versionValue.setText("2.0")
+
         # Add icon image
-        project_dir = os.path.dirname(os.path.dirname(
+        base_dir = os.path.dirname(os.path.dirname(
             os.path.abspath(__file__)))  # Get parent directory
-        pixmap = QtGui.QPixmap(os.path.join(project_dir, 'images', 'icon.png'))
+        pixmap = QtGui.QPixmap(os.path.join(base_dir, 'images', 'icon.png'))
         self.icon.setPixmap(pixmap)
 
         # Delete the object when dialog is closed